backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-multipart==0.0.12
aiohttp==3.10.10
